+++ crud/templates/functions.php
@@ -49,26 +49,32 @@ function generateReport()
     $serializedData = file_get_contents(DB_NAME);
     $students = unserialize($serializedData);
     ?>
-        <table>
+    <table>
+        <tr>
+            <th>ID</th>
+            <th>Name</th>
+            <th>Roll</th>
+            <?php if (isAdmin() || isEditor()): ?>
+            <th width="25%">Action</th>
+            <?php endif; ?>
+        </tr>
+        <?php
+        foreach ($students as $student) {
+            ?>
             <tr>
-                <th>ID</th>
-                <th>Name</th>
-                <th>Roll</th>
-                <th width="25%">Action</th>
+                <td><?php printf('%d', $student['id']) ?></td>
+                <td><?php printf('%s %s', $student['fname'], $student['lname']); ?></td>
+                <td><?php printf('%s', $student['roll']); ?></td>
+                <?php if (isAdmin()): ?>
+                    <td><?php printf('<a href="/crud/index.php?task=edit&id=%s">Edit</a> | <a class="delete" href="/crud/index.php?task=delete&id=%s">Delete</a>', $student['id'], $student['id']); ?></td>
+                <?php elseif (isEditor()): ?>
+                    <td><?php printf('<a href="/crud/index.php?task=edit&id=%s">Edit</a>',$student['id']); ?></td>
+                <?php endif; ?>
             </tr>
             <?php
-foreach ($students as $student) {
+        }
         ?>
-                <tr>
-                    <td><?php printf('%d', $student['id'])?></td>
-                    <td><?php printf('%s %s', $student['fname'], $student['lname']);?></td>
-                    <td><?php printf('%s', $student['roll']);?></td>
-                    <td><?php printf('<a href="/crud/index.php?task=edit&id=%s">Edit</a> | <a class="delete" href="/crud/index.php?task=delete&id=%s">Delete</a>', $student['id'], $student['id']);?></td>
-                </tr>
-                <?php
-}
-    ?>
-        </table>
+    </table>
     <?php
 }
 
@@ -133,28 +139,42 @@ function updateStudent($id, $fname, $lname, $roll)
     return false;
 }
 
-function deleteStudent($id){
+function deleteStudent($id)
+{
     $serializedData = file_get_contents(DB_NAME);
     $students = unserialize($serializedData);
 
-    foreach($students as $offset=>$student){
-        if($student['id']==$id){
+    foreach ($students as $offset => $student) {
+        if ($student['id'] == $id) {
             unset($students[$offset]);
         }
     }
     $serializedData = serialize($students);
-    file_put_contents(DB_NAME,$serializedData,LOCK_EX);
+    file_put_contents(DB_NAME, $serializedData, LOCK_EX);
 }
 
-function printRaw(){
+function printRaw()
+{
     $serializedData = file_get_contents(DB_NAME);
-    $students =     unserialize($serializedData);
+    $students = unserialize($serializedData);
     print_r($students);
 }
 
-function getNewId($students){
-    $maxId = max(array_column($students,'id'));
-    return $maxId+1;    
+function getNewId($students)
+{
+    $maxId = max(array_column($students, 'id'));
+    return $maxId + 1;
 }
 
-?>
+function isAdmin()
+{
+    return ('admin' == $_SESSION['role']);
+}
+
+function isEditor()
+{
+    return ('editor' == $_SESSION['role']);
+}
+function hasPrivilege(){
+    return (isAdmin() || isEditor());
+}
